backend/pyproject.toml
[tool.poetry]
name = "iwallpaper"
version = "0.0.0"
description = "Wallpaper exploring platform."
authors = ["Yuwang Cai <mrcaidev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
aiodns = "^3.0.0"
aiofiles = "^23.1.0"
aiohttp = "^3.8.4"
fastapi = "^0.100.0"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
scikit-learn = "^1.3.0"
supabase = "^1.0.3"
torch = "^2.0.1"
torchvision = "^0.15.2"
uvicorn = "^0.23.0"
vecs = "^0.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
